+++ src/lib/DateInput.svelte
@@ -84,6 +84,9 @@
 
   /** Whether the date popup is visible */
   export let visible = false
+  /** Close the date popup when a date is selected */
+  export let closeOnSelection = false
+
   // handle on:focusout for parent element. If the parent element loses
   // focus (e.g input element), visible is set to false
   function onFocusOut(e: FocusEvent) {
@@ -113,6 +116,9 @@
 
   function onSelect(e: CustomEvent<undefined>) {
     dispatch('select', e.detail)
+    if (closeOnSelection) {
+      visible = false
+    }
   }
 </script>